Added figures relating to community growth, (daily / weekly / monthly).
One of the pieces of feedback that we got for this notebook was that
the data would be better interpretable if it was in a more granular
view than aggregate. This commit updates the notebook to meet this suggestion.

For each analyzed category, plots are added to show changes by time-bin,
e.g. (daily, weekly, monthly).

Signed-off-by: JamesKunstle <jkunstle@redhat.com>